refactored deploy, added update and Pep8
