+++ src/viur_cli/__init__.py
@@ -9,3 +9,4 @@ from .setup import *
 from .version import *
 from .scriptor import script
 from .tool import *
+from .update import *
