+++ src/viur_cli/deploy.py
@@ -1,5 +1,9 @@
-import click, os, string, sys, re, yaml
+import click
+import os
+import string
+import yaml
 from . import cli, echo_error, get_config, echo_info, replace_vars
+from .update import create_req
 
 
 @cli.command(context_settings={"ignore_unknown_options": True})
@@ -33,10 +37,11 @@ def deploy(action, name, additional_args):
         version = "".join([c for c in version.lower() if c in string.ascii_lowercase + string.digits + "-"])
 
         # rebuild requirements.txt
-        create_req()
+        create_req(False)
 
         os.system(
-            f'gcloud app deploy --project={conf["application_name"]} --version={version} --no-promote {" ".join(additional_args)} {conf["distribution_folder"]}')
+            f'gcloud app deploy --project={conf["application_name"]} --version={version} '
+            f'--no-promote {" ".join(additional_args)} {conf["distribution_folder"]}')
     else:
         if action not in ["index", "queue", "cron"]:
             echo_error(f"{action} is not a valid action. Valid is app, index, queue, cron")
@@ -49,8 +54,8 @@ def deploy(action, name, additional_args):
                 with open(yaml_file, "r") as source_file:
                     try:
                         data = yaml.safe_load(source_file)
-                    except:
-                        raise ValueError()
+                    except ValueError as e:
+                        raise e
 
                     if "indexes" not in data:
                         raise ValueError()
@@ -82,89 +87,3 @@ def deploy(action, name, additional_args):
 
         os.system(
             f'gcloud app deploy --project={conf["application_name"]} {" ".join(additional_args)} {yaml_file}')
-
-def create_req():
-    """
-    load projects pipenv and build a requirements.txt
-
-    cores requirements.txt cant be validated currently, because of the core does not provide a feature for that
-    """
-    projectConfig = get_config()
-    distFolder = projectConfig["default"]["distribution_folder"]
-    if projectConfig["default"]["core"] != "submodule":
-
-        if click.confirm(f"Do you want to regenerate the requirements.txt located in the {distFolder}?"):
-            os.system(f"pipfile2req  --hashes > {distFolder}/requirements.txt")
-
-            file_object = open(f"{distFolder}/requirements.txt", 'r')
-            generated_requirements = file_object.read()
-            for line in generated_requirements.splitlines():
-                if "]==" in line:
-                    # we got a dependency with extras
-                    generated_requirements+=re.sub(r"\[.*?\]","",line)+"\n"
-            file_object.close()
-
-            file_obj = open(f"{distFolder}/requirements.txt", 'w')
-            file_obj.write(generated_requirements)
-            file_obj.close()
-            echo_info("requirements.txt successfully generated")
-
-        if check_req(f"{distFolder}/requirements.txt"):
-            if not click.confirm(f"There are some depencency errors, are you sure you want to continue?"):
-                sys.exit(0)
-
-
-def check_req(projects_requirements_path):
-    import site
-    from pip._internal.req import parse_requirements
-    from pip._internal.network.session import PipSession
-
-    def requirements_to_dict(requirements):
-
-        ret = {}
-        for requirement in requirements:
-            package, version = requirement.requirement.split(";")[0].split("==")
-            package = package.split("[")[0]
-
-            requirement.options.update({"version":version.strip()})
-            ret.update({package.lower():requirement.options})
-        return ret
-
-    sp = site.getsitepackages()[0]
-    core_requirements = None
-    for req in (
-        os.path.join(sp, "viur", "core", "requirements.txt"),
-        os.path.join(sp, "viur", "requirements.txt")
-    ):
-        if os.path.exists(req):
-            core_requirements = req
-            break
-
-    if not core_requirements:
-        echo_error("could now find core package, please update the core to validate the requirements.txt")
-
-    core_requirements_obj = requirements_to_dict(parse_requirements(core_requirements, session=PipSession()))
-
-    projects_requirements_obj = requirements_to_dict(parse_requirements(projects_requirements_path, session=PipSession()))
-
-
-    errors = []
-    for package,options in core_requirements_obj.items():
-        if package not in projects_requirements_obj:
-            errors.append(f"missing package: {package} with version {options['version']}")
-            continue
-        elif options["version"]!=projects_requirements_obj[package]["version"]:
-            errors.append(f"version mismatch: expected {options['version']} got {projects_requirements_obj[package]['version']}: {package}")
-            continue
-        else:
-            # package exists, test hash
-            project_hashes = projects_requirements_obj[package]["hashes"]["sha256"]
-            core_hashes = options["hashes"]["sha256"]
-
-            if not set(core_hashes).issubset(set(project_hashes)):
-                errors.append(f"package hash mismatch: {package}")
-
-    for error in errors:
-        echo_error(error)
-
-    return errors
